% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pharmax.viz-package.R
\docType{package}
\name{pharmax.viz-package}
\alias{pharmax.viz}
\alias{pharmax.viz-package}
\title{pharmax.viz: Pharmacometric Visualization}
\description{
Publication-ready pharmacometric visualization helpers for the pharmax
ecosystem.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/pharmax-org/pharmax}
  \item \url{https://pharmax-org.github.io/pharmax}
  \item Report bugs at \url{https://github.com/pharmax-org/pharmax/issues}
}

}
\author{
\strong{Maintainer}: Malek Okour \email{malek@zajaly.com}

}
\keyword{internal}
