% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pharmax.R
\name{px_colors}
\alias{px_colors}
\title{pharmax color palette}
\usage{
px_colors(n = 8)
}
\arguments{
\item{n}{Number of colors.}
}
\value{
Character vector of hex colors.
}
\description{
pharmax color palette
}
