% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_npde.R
\name{px_npde}
\alias{px_npde}
\title{NPDE diagnostic panel}
\usage{
px_npde(
  data,
  npde = "NPDE",
  time = "TIME",
  prediction = "PRED",
  mode = "light"
)
}
\arguments{
\item{data}{Data frame containing an NPDE-like column.}

\item{npde}{NPDE column name.}

\item{time}{Optional time column.}

\item{prediction}{Optional prediction column.}

\item{mode}{Theme mode passed to \code{\link[=theme_pharmax]{theme_pharmax()}}.}
}
\value{
A patchwork plot object.
}
\description{
NPDE diagnostic panel
}
