% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_plot_grid.R
\name{px_plot_grid}
\alias{px_plot_grid}
\title{Compose pharmax plots}
\usage{
px_plot_grid(
  ...,
  ncol = 2,
  title = NULL,
  subtitle = NULL,
  caption = NULL,
  mode = "light"
)
}
\arguments{
\item{...}{ggplot2 or patchwork plots, or a single list of plots.}

\item{ncol}{Number of columns.}

\item{title}{Optional grid title.}

\item{subtitle}{Optional grid subtitle.}

\item{caption}{Optional grid caption.}

\item{mode}{Theme mode passed to \code{\link[=theme_pharmax]{theme_pharmax()}}.}
}
\value{
A patchwork plot object.
}
\description{
Compose pharmax plots
}
