% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/px_save.R
\name{px_save}
\alias{px_save}
\title{Save a pharmax plot}
\usage{
px_save(plot, filename, width = 8, height = 6, dpi = 300, ...)
}
\arguments{
\item{plot}{A ggplot2 or patchwork plot.}

\item{filename}{Output filename.}

\item{width}{Width in inches.}

\item{height}{Height in inches.}

\item{dpi}{Resolution.}

\item{...}{Additional arguments passed to \code{\link[ggplot2:ggsave]{ggplot2::ggsave()}}.}
}
\value{
The input filename, invisibly.
}
\description{
Save a pharmax plot
}
