% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_pharmax.R
\name{scale_color_pharmax}
\alias{scale_color_pharmax}
\alias{scale_fill_pharmax}
\title{pharmax color scale for ggplot2}
\usage{
scale_color_pharmax(...)

scale_fill_pharmax(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[ggplot2:discrete_scale]{ggplot2::discrete_scale()}}.}
}
\description{
pharmax color scale for ggplot2
}
